Changing Storage API key from URI to String (#1051)
Motivation: String is more generic and gives flexibility to the storage implementation to structure their own keys.
For example, the storage layer may use a JSON string as an identifier. This is not currently a use case, but going forward it would probably be better to not impose an unnecessary restriction on a top level API.

Tested separately on dev clusters.