Guicing MetricsManager, CommonMetrics, and WebMetrics and handling related changes (#1189)
Initial motivation was to remove ImmutableEnumChecker warnings. Worked with @kunkun-tang to determine that Guicing the offending classes was the best way to proceed.

MetricsManager, CommonMetrics, and WebMetrics had to be guicified together in order to handle tests that used both CommonMetrics/MetricsManager or WebMetrics/MetricsManager.

A number of tests required minor refactoring to handle the new guicier versions.

Certain classes were only partially guiced in order to minimize the size of this PR. TODOs were added so these classes could be followed up on and fully guiced in the future.

I'm hoping these testing changes/guicing overall will allow for more parallel testing. Additional adjustment of testing classes is likely necessary to make that happen.