Refactor: Guice injection of Azkaban Web Server dependencies (#1255)
The current code uses the `SERVICE_PROVIDER` interface to fetch dependencies for the Web Server. Refactored the code to enable parameterized dependency injection from the construction which is the standard way of declaring dependencies. In the process, removed legacy constructor from `AzkabanWebServer` and removed legacy class `AzkabanServletContextListener` which is no longer being used.

Injection sanity tests already exist in the code. Tested solo server and it works as well.