Creating AzkabanProjectLoader class to handle project upload/download (#1272)
The current upload logic is sitting in the ProjectManager class. This change delegates the upload and download to AzkabanProjectLoader class. This makes the upload/download code much easier to manage and test. Added a unit test for upload and download methods.

This change also helps pave the way for subsequent refactors which are required to clean up and simplify the project loading and validating logic.

Note: There is no logic change in this code.