fix guice initialize TriggerManager twice bug (#1283)
This bug was found in local solo server built by July 19th's master
code. Our integration test tried to add trigger, which run every
minutes. The strange observation is that this trigger never run. We
debugged the code and found that the TriggerScannerThread was launched
twice, and newly added triggers were put into the should-not-exist
thread and never run. I noticed that TriggerManager did not bind to
Singleton in Guice Confs. I made this change, and the thread doesn't
launch twice. The bug disappeared.